modelit_test/requirements.txt
torch>=2.0.0
numpy>=1.23.2
Pillow>=9.0.0
tqdm>=4.65.0
psutil>=5.9.0
pandas>=2.0.0
pyarrow>=12.0.1
loguru>=0.7.0
fsspec>=2022.11.0
gputil==1.4.0
datasets>=2.15.0
huggingface_hub>=0.24.0
requests>=2.31.0
pyyaml>=6.0
matplotlib>=3.7.0
ipython>=8.10.0
open3d>=0.16.0 